Removed sort from exhibit related items.
Previously sorting on image to group images first and then articles, but with
the new styles for related items it is preferable for users to select the order
in which related items appear. By removing all sorts it should default to the
order in which they are entered, if not we might need to add field collection
delta or something.